Formalize the lifecycle of Exchange
There's four useful events:

1. When we're ready to call network interceptors. At this point
   we expect an exchange to exist, as a handle to the connection
   we've chosen for the network call

2. When we hold an exchange that must later be released.

3. When we release the exchange

4. When we're done calling network interceptors. Only at this point
   do we no longer need to remember what the exchange was and what
   connection it used.

In practice 1 and 2 happen mostly at the same time. Right now 1 is
in RetryAndFollowUpInterceptor and 2 is in ConnectInterceptor, but
we should be able to move those around without pain.

But we have a problem with 3 and 4. If the response has no response
body, then 3 will precede 4. We will release the exchange as soon as
we know there's no streams held. This means that network interceptors,
including our own RetryAndFollowUpInterceptor, do not consistently
have an exchange after they've called proceed(). This is problematic
if they want to make decisions based on what connection was used.

I've added a new member, 'interceptorScopedExchange' that holds the
exchange from 1 and 2 through 4, regardless of when 3 happens.

This is all working towards a solution to handling HTTP 421s on
coalesced connections.

https://github.com/square/okhttp/issues/5424